fix: eas-cli build android run gradle エラーに対応 #73
```
Check the underlying message for the error that occurred while compiling the expo-ads-admob. It may be that this library is not compatible with your version of React Native, or there may be another issue with an underlying dependency version. Check Expo Doctor for warnings and the Github repo for this library to see if the version you are using is compatible or if there are any open issues or bug fixes. Note that this is the Gradle package name of the library, and that may not entirely match the name of the NPM package or Github repo.
```